+++ std-bits/table/src/main/java/org/enso/table/parsing/BaseTimeParser.java
@@ -7,7 +7,7 @@ import org.enso.table.data.column.builder.object.Builder;
 import org.enso.table.data.column.builder.object.ObjectBuilder;
 import org.enso.table.parsing.problems.InvalidFormatProblemAggregator;
 
-public abstract class BaseTimeParser extends TypeParser<InvalidFormatProblemAggregator> {
+public abstract class BaseTimeParser extends DatatypeParser<InvalidFormatProblemAggregator> {
   protected interface ParseStrategy {
     Object parse(String text, DateTimeFormatter formatter) throws DateTimeParseException;
   }
