+++ std-bits/table/src/main/java/org/enso/table/parsing/problems/AdditionalInvalidRows.java
@@ -1,10 +1,4 @@
 package org.enso.table.parsing.problems;
 
 /** A problem which indicates how many additional invalid rows were encountered. */
-public class AdditionalInvalidRows implements ParsingProblem {
-  public final long count;
-
-  public AdditionalInvalidRows(long count) {
-    this.count = count;
-  }
-}
+public record AdditionalInvalidRows(long count) implements ParsingProblem {}
