+++ std-bits/table/src/main/java/org/enso/table/parsing/problems/InvalidFormat.java
@@ -2,4 +2,5 @@ package org.enso.table.parsing.problems;
 
 import java.util.List;
 
+/** Indicates that a text value did not match the format expected of a datatype. */
 public record InvalidFormat(List<String> cells) implements ParsingProblem {}
