+++ std-bits/table/src/main/java/org/enso/table/parsing/problems/InvalidFormatProblemAggregator.java
@@ -3,10 +3,19 @@ package org.enso.table.parsing.problems;
 import java.util.ArrayList;
 import java.util.List;
 
+/**
+ * A base problem aggregator that allows reporting the most generic {@code InvalidFormat} problem.
+ */
 public class InvalidFormatProblemAggregator implements ProblemAggregator {
 
   private final List<String> invalidFormatCells = new ArrayList<>();
 
+  /**
+   * Reports a cell with an invalid format.
+   *
+   * <p>The reports are aggregated and finally a single problem containing all invalid cell for the
+   * given column is reported.
+   */
   public void reportInvalidFormat(String cell) {
     invalidFormatCells.add(cell);
   }
