+++ std-bits/table/src/main/java/org/enso/table/parsing/problems/LeadingZeros.java
@@ -2,4 +2,5 @@ package org.enso.table.parsing.problems;
 
 import java.util.List;
 
+/** Indicates that some values contained leading zeros when leading zeros where not allowed in the given numeric conversion. */
 public record LeadingZeros(List<String> cells) implements ParsingProblem {}
